chore: review batch — residue compaction, float semantics, hygiene
- residue: fix _01forward_indexed double-compaction. partword from
  res1_class is already compacted to `used` rows; the function was
  re-indexing via `indices` which would OOB on stereo-with-silent-
  channel (returning empty Vec, then panicking on access). Latent
  bug — current corpus had nonzero on both channels.

- floor1::inspect_error: use float semantics matching C. Override
  comparisons (`y + maxover < val`) and squared-max-rate-vs-maxerr
  checks now run in f32. The mse/n check keeps int truncation
  (C does int/int, then promotes the truncated int to float for
  the comparison).

- encode: mode lookup .unwrap_or(0/1) → .expect (closed input space
  guarantees both modes exist); apply granule_pos cap only on the
  final packet, debug_assert non-final packets stay <= total_samples;
  random_serial mixes nanos + seconds + stack-addr for wider entropy;
  drop dead decoded_samples binding.

- tables/mod.rs: pub → pub(crate) (no public consumers).

corpus_parity_44_stereo + 9 oracle parity tests still pass.

+++ src/floor1.rs
@@ -615,11 +615,14 @@ fn inspect_error(
     mse = y - val;
     mse *= mse;
     n += 1;
+    // C: `if(y+info->maxover<val)` — int + float < int promotes both sides to
+    // float. For Q5 maxover/maxunder are integers (60/30) so the f32 path is
+    // bit-identical to the prior i32 cast; preserved as f32 for non-Q5 floors.
     if mdct[x as usize] + info.twofitatten >= mask[x as usize] {
-        if y + (info.maxover as i32) < val {
+        if (y as f32) + info.maxover < val as f32 {
             return 1;
         }
-        if y - (info.maxunder as i32) > val {
+        if (y as f32) - info.maxunder > val as f32 {
             return 1;
         }
     }
@@ -642,23 +645,27 @@ fn inspect_error(
         n += 1;
         if mdct[x as usize] + info.twofitatten >= mask[x as usize] {
             if val != 0 {
-                if y + (info.maxover as i32) < val {
+                if (y as f32) + info.maxover < val as f32 {
                     return 1;
                 }
-                if y - (info.maxunder as i32) > val {
+                if (y as f32) - info.maxunder > val as f32 {
                     return 1;
                 }
             }
         }
     }
 
-    if info.maxover as i32 * info.maxover as i32 / n > info.maxerr as i32 {
+    // C: `info->maxover*info->maxover/n > info->maxerr` — float*float / int
+    // is f32 (int promotes to f32 for the division). Compared to f32 maxerr.
+    if info.maxover * info.maxover / n as f32 > info.maxerr {
         return 0;
     }
-    if info.maxunder as i32 * info.maxunder as i32 / n > info.maxerr as i32 {
+    if info.maxunder * info.maxunder / n as f32 > info.maxerr {
         return 0;
     }
-    if mse / n > info.maxerr as i32 {
+    // C: `mse/n > info->maxerr` — int/int (truncates), then int > float
+    // promotes the truncated int to f32. Keep the integer truncation.
+    if ((mse / n) as f32) > info.maxerr {
         return 1;
     }
     0
